# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Webstore Customer API",
        "description": "Dokumentasi endpoint API customer untuk autentikasi dan fitur terkait dashboard.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Default API server"
        }
    ],
    "paths": {
...
        "/api/auth/register-meta": {
            "get": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Metadata halaman registrasi customer",
                "description": "Mengembalikan metadata SEO dan referral code aktif yang digunakan pada halaman registrasi web customer.",
                "operationId": "84b3a14f397fe6fe9c7162b99be9fc9e",
                "parameters": [
                    {
                        "name": "referral_code",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadata registrasi berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "referralCode": "ABCD1234",
                                        "seo": {
                                            "title": "Daftar Jadi Member",
                                            "description": "Bergabunglah sebagai member Webstore.",
                                            "canonical": "http://localhost/register"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Registrasi customer API",
                "description": "Mendaftarkan customer baru melalui API sesuai rule registrasi customer web.",
                "operationId": "c7bb21a0d87049e161b8e4c9b0d909f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "username",
                                    "email",
                                    "telp",
                                    "gender",
                                    "password",
                                    "password_confirmation",
                                    "terms"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "budi01"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "budi@example.com"
                                    },
                                    "telp": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "nik": {
                                        "type": "string",
                                        "example": "3276010101010001",
                                        "nullable": true
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "L"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 1",
                                        "nullable": true
                                    },
                                    "referral_code": {
                                        "type": "string",
                                        "example": "ABCD1234",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret12345"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret12345"
                                    },
                                    "terms": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Registrasi customer berhasil.",
                                    "data": {
                                        "customer": {
                                            "id": 1,
                                            "name": "Budi Santoso",
                                            "username": "budi01",
                                            "email": "budi@example.com",
                                            "phone": "08123456789",
                                            "status": 1,
                                            "ref_code": "ABCD1234"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Data tidak valid.",
                                    "errors": {
                                        "email": [
                                            "Email sudah terdaftar."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Login customer API",
                "description": "Autentikasi customer menggunakan username dan password, lalu generate Sanctum personal access token.",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "member001"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "android-app",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login API berhasil."
                                        },
                                        "data": {
                                            "properties": {
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "1|abcdefghijklmnopqrstuvwxyz"
                                                },
                                                "customer": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Budi Santoso"
                                                        },
                                                        "username": {
                                                            "type": "string",
                                                            "example": "budi01"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "budi@example.com"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "08123456789",
                                                            "nullable": true
                                                        },
                                                        "status": {
                                                            "type": "integer",
                                                            "example": 3
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Login API berhasil.",
                                    "data": {
                                        "token_type": "Bearer",
                                        "access_token": "1|abcdefghijklmnopqrstuvwxyz",
                                        "customer": {
                                            "id": 1,
                                            "name": "Budi Santoso",
                                            "username": "budi01",
                                            "email": "budi@example.com",
                                            "phone": "08123456789",
                                            "status": 3
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal atau kredensial tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Username atau kata sandi salah.",
                                    "errors": {
                                        "username": [
                                            "Username atau kata sandi salah."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Logout customer API",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logout berhasil",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Logout API berhasil."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak terautentikasi."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dengan format: Bearer {token}",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}